health-bot/requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
APScheduler==3.10.4
